requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
faker==20.1.0
groq==0.4.1
pillow==10.1.0
python-dotenv==1.0.0
Flask==3.0.3
flask-cors==4.0.0
openai==0.27.8
logging==0.5.1